+++ packages/react-ui/src/app/components/dashboard-page-header.tsx
@@ -1,16 +1,13 @@
-import { BetaBadge } from '@/components/custom/beta-badge';
 import { useEmbedding } from '@/components/embed-provider';
 
 export const DashboardPageHeader = ({
   title,
   children,
   description,
-  beta = false,
 }: {
   title: string;
   children?: React.ReactNode;
   description?: React.ReactNode;
-  beta?: boolean;
 }) => {
   const { embedState } = useEmbedding();
 
@@ -23,11 +20,6 @@ export const DashboardPageHeader = ({
         <div>
           <div className="flex items-center gap-2">
             <h1 className="text-xl font-semibold">{title}</h1>
-            {beta && (
-              <div className="flex items-center">
-                <BetaBadge />
-              </div>
-            )}
           </div>
           {description && (
             <span className="text-xs text-muted-foreground">{description}</span>
