+++ packages/react-ui/src/app/components/project-settings/environment/index.tsx
@@ -6,9 +6,9 @@ import LockedFeatureGuard from '@/app/components/locked-feature-guard';
 import { Button } from '@/components/ui/button';
 import { Card } from '@/components/ui/card';
 import { LoadingSpinner } from '@/components/ui/spinner';
-import { ConnectGitDialog } from '@/features/git-sync/components/connect-git-dialog';
-import { gitSyncApi } from '@/features/git-sync/lib/git-sync-api';
-import { gitSyncHooks } from '@/features/git-sync/lib/git-sync-hooks';
+import { ConnectGitDialog } from '@/features/project-releases/components/connect-git-dialog';
+import { gitSyncApi } from '@/features/project-releases/lib/git-sync-api';
+import { gitSyncHooks } from '@/features/project-releases/lib/git-sync-hooks';
 import { platformHooks } from '@/hooks/platform-hooks';
 import { authenticationSession } from '@/lib/authentication-session';
 import { assertNotNullOrUndefined } from '@activepieces/shared';
