+++ packages/react-ui/src/app/guards/permission-guard.tsx
