+++ packages/react-ui/src/app/routes/embed/index.tsx
@@ -4,7 +4,7 @@ import { flushSync } from 'react-dom';
 import { useTranslation } from 'react-i18next';
 import { useEffectOnce } from 'react-use';
 
-import { memoryRouter } from '@/app/router';
+import { memoryRouter } from '@/app/guards';
 import { useEmbedding } from '@/components/embed-provider';
 import { useTheme } from '@/components/theme-provider';
 import { LoadingScreen } from '@/components/ui/loading-screen';
