+++ packages/react-ui/src/app/routes/project-release/create-release-dialog/index.tsx
@@ -20,8 +20,8 @@ import { Label } from '@/components/ui/label';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { LoadingSpinner } from '@/components/ui/spinner';
 import { Textarea } from '@/components/ui/textarea';
-import { gitSyncHooks } from '@/features/git-sync/lib/git-sync-hooks';
-import { projectReleaseApi } from '@/features/project-version/lib/project-release-api';
+import { gitSyncHooks } from '@/features/project-releases/lib/git-sync-hooks';
+import { projectReleaseApi } from '@/features/project-releases/lib/project-release-api';
 import { platformHooks } from '@/hooks/platform-hooks';
 import { authenticationSession } from '@/lib/authentication-session';
 import {
