+++ packages/react-ui/src/app/routes/project-release/push-everything-dialog/index.tsx
@@ -28,8 +28,8 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from '@/components/ui/tooltip';
-import { gitSyncApi } from '@/features/git-sync/lib/git-sync-api';
-import { gitSyncHooks } from '@/features/git-sync/lib/git-sync-hooks';
+import { gitSyncApi } from '@/features/project-releases/lib/git-sync-api';
+import { gitSyncHooks } from '@/features/project-releases/lib/git-sync-hooks';
 import { platformHooks } from '@/hooks/platform-hooks';
 import { authenticationSession } from '@/lib/authentication-session';
 import {
