+++ packages/react-ui/src/app/routes/project-release/selection-release-dialog/project-dialog.tsx
@@ -17,7 +17,7 @@ import {
 } from '@/components/ui/dialog';
 import { FormField, FormItem, Form, FormMessage } from '@/components/ui/form';
 import { Label } from '@/components/ui/label';
-import { projectReleaseApi } from '@/features/project-version/lib/project-release-api';
+import { projectReleaseApi } from '@/features/project-releases/lib/project-release-api';
 import { projectHooks } from '@/hooks/project-hooks';
 import { DiffReleaseRequest, ProjectReleaseType } from '@activepieces/shared';
